pyproject.toml
[tool.poetry]
name = "uzen"
version = "0.1.0"
description = "YARA with Puppeteer"
authors = ["Manabu Niseki <manabu.niseki@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
aiocache = "^0.11.1"
aiofiles = "^0.8.0"
aiometer = "^0.3.0"
aiomysql = "^0.0.22"
aioredis = "1.3.1"
aiosqlite = "^0.17.0"
arq = "^0.22"
beautifulsoup4 = "^4.10.0"
cached-property = "^1.5.2"
certifi = "^2021.10.8"
dateparser = "^1.1.0"
dnspython = "^2.2.0"
fastapi = "^0.73.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.3"}
gunicorn = "^20.1.0"
html2text = "^2020.1.16"
httpx = "^0.21.3"
loguru = "^0.6.0"
minio = "^7.1.3"
niteru = "^0.2.1"
playwright = "^1.18.2"
playwright-har-tracer = "^0.3.2"
ppdeep = "^20200505"
pyhumps = "^3.5.3"
pyopenssl = "^22.0.0"
python-ulid = "^1.0.3"
sentry-sdk = "^1.5.4"
tldextract = "^3.1.2"
tortoise-orm = "^0.18.1"
ujson = "^5.1.0"
uvicorn = {extras = ["standard"], version = "^0.17.4"}
whois-parser = "^0.1.4"
yara-python = "^4.1.3"

[tool.poetry.dev-dependencies]
async-timeout = "3.0.1"
autoflake = "^1.4"
autopep8 = "^1.6.0"
black = "^22.1.0"
coveralls = "^3.3.1"
flake8 = "^4.0.1"
isort = "^5.10.1"
mypy = "^0.931"
nest-asyncio = "^1.5.4"
pre-commit = "^2.17.0"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
pytest-black = "^0.3.12"
pytest-cov = "^3.0.0"
pytest-env = "^0.6.2"
pytest-mock = "^3.7.0"
pytest-parallel = "^0.1.1"
pytest-randomly = "^3.11.0"
pytest-sugar = "^0.9.4"
pytest-timeout = "^2.1.0"
pyupgrade = "^2.31.0"
requests = "^2.27.1"
respx = "^0.19.2"
vcrpy = "^4.1.1"

[tool.taskipy.tasks]
app = "uvicorn app:app --reload"
worker = "arq app.arq.worker.WorkerSettings --watch ./"

[tool.isort]
force_grid_wrap = 0
include_trailing_comma = true
known_third_party = ["OpenSSL", "aiometer", "alembic", "bs4", "dns", "fastapi", "fastapi_utils", "httpx", "ipwhois", "loguru", "playwright", "pydantic", "pyppeteer", "pysafebrowsing", "pytest", "requests", "respx", "sqlalchemy", "starlette", "tortoise", "vcr", "whois", "yara"]
line_length = 88
multi_line_output = 3
use_parentheses= true

[tool.pytest.ini_options]
env = ["IP2ASN_WEB_SERVICE_URL=http://127.0.0.1:80", "TESTING=True", "workers=auto"]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
